requirements.txt
SQLAlchemy==2.0.24
newspaper3k==0.2.8
python-dotenv==1.0.0
pytest==7.4.3
beautifulsoup4==4.12.2
pymysql
sqlalchemy_utils